Several changes in the class SparsePolynomial
Removed unused imports

Fixed some typos on documentation

Added methods monomials and coefficients that returns the actual monomials (as SparsePolynomials) and the coefficients (as elements of the domain)

Added reverse computations for addition, multiplication and substraction

Changed the equality method to recognize identity with element in the domain (in case the polynomial are constants)

Added magic method __pow__ as a call to method exp. Now the syntax ** works

Added static method "monomial" to he class SparsePolynomial that takes a monomial in the format saved in the _data field and build the actual SparsePolynomial that it represents.