Splitted the creation of matrices
Now the process of creating the matrices from the Jacobian is completely included in the method "construct_matrices".

This method had added two new parameters: one deciding whether or not to use random evaluation or the usual algorithms. The second only affects the latter, and remove linearly dependent generated matrices.

The method "construct_matrices_evaluation_random"generates a set of lienarly independent matrices evaluating the jacobian matrix.
It is a loop of evaluations until a linearly dependen matrix appears. Then we use Schwarz-Zippel Lemma.

Here we have not use the full power of this Lemma and we have a very high probability of having a basis of the vector space.

Methods "do_lumping" and "do_lumping_internal" now include a new optional argument to use the random evaluations or not.

In "do_lumping_internal", we changed the part of the code to make the creation of matrices go only through the method "construct_matrices".

Checked with the example BIOMD0000000033. Last execution was 6 minutes.