Fixed the simplify method
This method was raising some errors when the denominator had no variables due to some internal conversion in sympy.

The method __floordiv__ of SparsePolynomial has been adapted by avoiding going to Sympy until it is completely necessary. This fixes the issue.

There are further checkings to be added since the gcd procedure does not work properly (detecting the constant gcd) when the coefficients are rational numbers.