Added lumping to class for systems
small rearrange of import

Changed system for the logger: now it works internally in the module.

added properties size and type to class FODESystem

fixed output of method construct_matrices

added method lumping (equivalent to the external "do_lumping")
added method _lumping(equivalent to the external "do_lumping_internal")

Created a new class "LODESystem"to represent lumped systems.
This class includes a new attribute that is the expressions of the new variables in term of the old variables.