Improved the interface of systems
Added method to get the equation for a variable from its name
Added more information to lumped systems (a pointer to the old system)
Added method "check_consistency" to plug random values between systems and check if they provide the same information (this is the idea for testing quickly if a lumping is correct).
Added method "is_consistent" to lumped systems.

Changed name of lumped systems to LDESystem.