Increase functinality of check_consistency
Added an optional parameter to the method check_consistency
that allows to symbolically check the consistency of a system.

This is terribly slow for rational systems, and an improvement when we are
handling sympy expressions may be useful.

Extended also the equality of Sparsepolynomials: now it detects the case of RationalFunction