Fixed error in CLUE
I realized that PolyElement may contain negative exponents, that lead to some errors using automatic differentiation.

I added a funciton "normalize" to the differential systems to unify the treatment of the different inputs throughout the code. Also, in the case that one of those Laurent polynomials appear, the sympy expression structure will be used.