Fixed method for computing digits
This method fails now in Python due to a change in the system: now there is a maximum on the number of digits of a int when casting it to a string.

Hence, in order to compute the number of digits we use the logarithm in base 10 and the ceil function.