Added method "find_in" in Subspace
This method computes the expression of a vector in terms of the subspace. The implementation is different between the Gaussian subspace and the Orthogonal subspace.

Changed in the exact_test the simulation: now we only simulate the exact lumping instead of the original system.

Adapted some times in the numerical examples